+++ bbs/sublist.cpp
@@ -248,7 +248,7 @@ void SubList() {
           bout.bprintf("|#1Select |#9[|#2%d-%d, [N]ext Page, [Q]uit|#9]|#0 : ", firstp + 1, lastp + 1);
           const std::string ss = mmkey(MMKeyAreaType::subs, true);
           if (isdigit(ss[0])) {
-            for (size_t i2 = 0; i2 < a()->subs().subs().size(); i2++) {
+            for (uint16_t i2 = 0; i2 < a()->subs().subs().size(); i2++) {
               if (ss == a()->usub[i2].keys) {
                 a()->set_current_user_sub_num(i2);
                 oldSub = a()->current_user_sub().subnum;
